requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.1
google-genai>=1.52.0
openai>=1.0.0
pydantic>=2.9.0
pillow>=12.0.0
python-pptx>=1.0.0
python-dotenv>=1.0.1
reportlab>=4.1.0
werkzeug>=3.0.1
markitdown[all]>=0.0.1
tenacity>=9.0.0
alembic>=1.13.0
flask-migrate>=4.0.0
img2pdf>=0.5.1
lazyllm[online-advanced]>=0.7.3
zhipuai>=2.1.5.20250708
volcengine-python-sdk[ark]>=5.0.9
PyPDF2>=3.0.0
PyMuPDF>=1.24.0
dashscope>=1.25.13
